server/requirements.txt
geopy==1.11.0
nose==1.3.7
Paste==2.0.3
python-dateutil==2.5.3
python-gflags==2.0
six==1.10.0
webapp2==2.5.2
WebOb==1.6.1
